edge_ai_innovations/requirements.txt
tensorflow>=2.13.0
tensorflow-datasets>=4.9.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
jupyter>=1.0.0
notebook>=7.0.0
scikit-learn>=1.3.0

# For Raspberry Pi deployment (uncomment when deploying to Pi)
# tflite-runtime>=2.13.0
